app/crypto-perf: update result granularity
In cases like RSA which has low performance numbers
compared to AES, the Mops and gbps may not give accurate results.
Hence updated the number of decimal places to have meaningful
results.
The results are not updated to show Kops or mbps as the results
may be used in CI scripts which considers them as Mops/gbps and
hence all scripts will need to be updated.

Signed-off-by: Akhil Goyal <gakhil@marvell.com>